+++ twoOpt.h
@@ -16,18 +16,18 @@
 using namespace std;
 
 // Non-looping version of two-opt optimization heuristic
-int twoOpt(int **graph, vector<int> &path, int &pathLength, int n);
+long long twoOpt(long long **graph, vector<long long> &path, long long &pathLength, long long n);
 
 // 2-Opt helper function: swap two nodes
-int is_path_shorter(int **graph, int v1, int v2, int v3, int v4, int &total_distance);
+long long is_path_shorter(long long **graph, long long v1, long long v2, long long v3, long long v4, long long &total_distance);
 
-void reverse(vector<int> &path, int start, int end, int n);
+void reverse(vector<long long> &path, long long start, long long end, long long n);
 
 
 
 
 // move this to tsp class
-int get_path_length(int **graph, vector<int> &path, int size);
+long long get_path_length(long long **graph, vector<long long> &path, long long size);
 
 
 #endif
